add new type of plag

+++ plagBetweenFiles.py
@@ -1,4 +1,5 @@
 import os
+
 import numpy as np
 from sklearn.feature_extraction.text import TfidfVectorizer
 from sklearn.metrics.pairwise import cosine_similarity
@@ -11,7 +12,6 @@ from sklearn.metrics.pairwise import cosine_similarity
 plagiarismResults = set()
 
 
-# Compute similarity between them.
 def vectorize(Data):
     return TfidfVectorizer().fit_transform(Data).toarray()
 
@@ -33,7 +33,6 @@ def cosineSimilarity(x, y):
 
 
 def PlagiarismChecker(filesWithTheirVectors):
-
     for curFile, curVector in filesWithTheirVectors:
         temp = filesWithTheirVectors.copy()
         current_index = temp.index((curFile, curVector))
@@ -41,7 +40,7 @@ def PlagiarismChecker(filesWithTheirVectors):
         for otherFile, otherVector in temp:
             scoreOfSimilarity = cosineSimilarity(curVector, otherVector)
             file_pair = sorted((curFile, otherFile))
-            score = (file_pair[0], file_pair[1], "{:.2%}".format( scoreOfSimilarity) ) 
+            score = (file_pair[0], file_pair[1], "{:.2%}".format(scoreOfSimilarity))
             plagiarismResults.add(score)
     return plagiarismResults
 
@@ -60,5 +59,6 @@ def solve():
 
 def deleteFiles():
     files = [doc for doc in os.listdir() if doc.endswith('.cpp')]
-    for f in files:
-        os.remove(f)
+    if len(files) > 0:
+        for f in files:
+            os.remove(f)
